Default bounds: Use precisely the right bounds, from the fields
Rather than trying to filter the generic parameters for ones used in
types, state the bounds in terms of the actual field types.

This is precisely correct.  It avoids lossage in situations where the
field type's implementation of the trait has bounds which depend in a
less-than-straightforward way on its generics.

There is still an issue with implementing traits which have
supertraits.  We ought to apply a `Self: Super` bound, but we don't
yet.