Abolish into_where_predicates_by_generic_parameters
Nothing calls this.  And, nothing should, because it would give wrong
answers.

Rename create_where_predicates_from_generic_parameters to
create_where_predicates_from_all_generic_parameters.  This will be
useful if we implement a `bound(...)` syntax that allows requesting
the "apply a bound for each generic parameter, regardless" mode.

If we don't want to implement such a mode, we should delete this
function, but let's not entangle this already large bugfix MR with
questions about future features and UI.